requirements.txt
Flask==2.3.3
flask-cors==4.0.0
cryptography==41.0.3
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
bcrypt==4.0.1
PyJWT==2.8.0
